+++ CHANGELOG.md
@@ -15,6 +15,9 @@
 
 ## 0.8.0
 
+* Enhance QueryOption.BROADCAST to use documents from owner nodes
+  For details please see: https://www.pivotaltracker.com/projects/1471320/stories/116412415
+
 * Upgrade netty from 4.1.0.CR3 to 4.1.0.CR7
 
 * Update JWT(Json Web Token) to use private key when SSL is enabled. Otherwise,